python-callgraph-error/requirements.txt
pyyaml==5.1
django==4.0.4
tensorflow-macos==2.5.0